chore(cosmos): refactor,  move base client
