+++ modules/cosmos/module.config.example.json
@@ -17,8 +17,10 @@
     },
     "modules": {
         "bank": {
-            "name": "bank",
             "account": "ethm1dakgyqjulg29m5fmv992g2y66m9g2mjn6hahwg"
+        },
+        "poa": {
+            "stakedAmount": "1000000"
         }
     }
 }
