+++ modules/cosmos/src/modules/PoA/module.ts
@@ -1,5 +1,3 @@
-import { CosmosModule } from "../../models/module";
-
-export interface PoAModule extends CosmosModule {
+export interface PoAModule {
     stakedAmount: string;
 }
