+++ modules/cosmos/test/modules/PoA/index.test.ts
@@ -2,16 +2,17 @@ import { expect } from "chai";
 import { describe, it, before } from "mocha";
 import { assertChainEnvironments, assertChainTypes } from "@testing/mocha/assertions";
 import { PoAClient } from "../../../src/modules/PoA/client";
-import { CosmosChain } from "../../../src/models/chain";
 import moduleConfig from "../../../module.config.example.json";
+import { Chain } from "@firewatch/core/chain";
 
 describe("PoA Module", () => {
     let poaClient: PoAClient;
-    const chain = moduleConfig.cosmos.chain;
+    const chain = moduleConfig.chain;
+    const poaModule = moduleConfig.modules.poa;
 
     before(async () => {
-        assertChainTypes(["cosmos"], chain as unknown as CosmosChain);
-        assertChainEnvironments(["localnet", "devnet", "testnet", "mainnet"], chain as unknown as CosmosChain);
+        assertChainTypes(["cosmos"], chain as unknown as Chain);
+        assertChainEnvironments(["localnet", "devnet", "testnet", "mainnet"], chain as unknown as Chain);
 
         poaClient = await PoAClient.connect(chain.urls.rpc);
     });
@@ -22,7 +23,7 @@ describe("PoA Module", () => {
             expect(validators.length).to.be.greaterThan(0);
 
             for (const validator of validators) {
-                expect(validator.tokens).to.equal(chain.module.stakedAmount);
+                expect(validator.tokens).to.equal(poaModule.stakedAmount);
             }
         });
         it("should have only self delegation for each validator", async function () {
